toolchain: fix libssp logic
On Solaris, the libc doesn't include ssp stuff, so we need GCC's libssp.
For some reason the libssp logic in this eclass is quite complex, and it
seems to assume a lot.  However, when libssp is not wanted, it makes the
configure script believe that the libc has the required symbols.  The
way want_ssp works, it actually always returns false, so for the sake of
platforms where ssp is not in the libc, we now return true when the libc
doesn't have ssp symbols, like on Solaris.

--HG--
extra : convert_revision : svn%3Adcf5e38c-c701-0410-9811-92a63bff3de1/trunk%4059844