eupdate: treat ChangeLogs special
Try to deal with additions to ChangeLog both in Prefix as well as gx86.
Merge new items on top of the existing log, regardless what is in there.
This comes in handy, in particulat with newer repomans which will
automatically update the ChangeLog file.

--HG--
extra : convert_revision : svn%3Adcf5e38c-c701-0410-9811-92a63bff3de1/trunk%4060088