Replaced profile manipulation in bootstrap_setup() with environment variables in bootstrap_stage3().
The current practice of storing bootstrap kludges in the profiles, to be
removed at the emerge --sync, makes the bootstrap unnecessarily fragile; an
emerge --sync at the wrong time can break everything, which means that if
you manage to break your portage tree in the middle of stage3 this is
basically unfixable. Moreover, it makes the bootstrap process more stateful
than it needs to be. (It is also problematic for unrelated reasons in the
rap bootstrap, which will be merged soon.)

This patch gets rid of the profile manipulation by replacing it by exporting
the relevant variables in stage3, making the bootstrap more robust.